Fix DB hookup and allow for Add/Post/Del requests to work
Fixed the key in the keys.js because it didn't include a database (new version of MongoDB Atlas has a different format then the tutorials I was going off of).
Then I fixed the files to allow it to function properly and tested with Postman, added a few users, deleted one, checked that the get request works.